fix(inference): un-clamp TPM refund; fail-fast on AuthError; Decimal budget error; cleanup
I1: TokenBucket.refund no longer clamps to capacity, preserving concurrent
callers' in-flight reservation deficits; _refill still caps its own additions.
Auth: AuthError added to never-catch tuple in _safe_agent_inference so a bad
API key propagates as ExceptionGroup instead of silently producing 100 PARSE_ERRORs.
M3: deleted dead _price() inner helper in estimate_run (defined, never called).
M4: BudgetExceededError.__init__ accepts Decimal; budget.py passes Decimal
directly; spent_usd/cap_usd attributes are Decimal throughout.
M1-doc: BudgetTrackingProvider.chat notes the bounded overshoot window.
Tests: 224 passed, 0 new mypy errors (baseline 34 unchanged).

+++ src/alphaswarm/errors.py
@@ -2,6 +2,8 @@
 
 from __future__ import annotations
 
+from decimal import Decimal
+
 
 # ---------------------------------------------------------------------------
 # Provider-agnostic inference errors
@@ -82,7 +84,7 @@ class BudgetExceededError(Exception):
         cap_usd: The configured budget cap in USD.
     """
 
-    def __init__(self, spent_usd: float, cap_usd: float) -> None:
+    def __init__(self, spent_usd: Decimal, cap_usd: Decimal) -> None:
         super().__init__(
             f"Budget cap exceeded: spent ${spent_usd:.4f} > cap ${cap_usd:.4f}"
         )
